Make sure 32-bit powerpc syscall defs don't leak to 64-bit powerpc systems
generate_syscalls.sh generates the syscalls.h header at the configure
time.  At the moment, that header has a set of 32-bit syscalls defined
with the __powerpc__ gate, plus another set of 64-bit syscalls defined
with the __powerpc64__ gate.  For 32-bit powerpc systems that's fine.
But for a 64-bit powerpc system this means that both sets of syscalls
become defined, which isn't right.

Thing is that on a 64-bit powerpc system, both __powerpc__ and
__powerpc64__ are defined compiler macros, while on a 32-bit powerpc
system, only the former is defined while the latter is not.

That said, the correct gate for a 32-bit only powerpc code is:
  #if defined(__powerpc__) && !defined(__powerpc64__)

Without this patch, e.g. __NR_clock_gettime64 def leaks to
64-bit powerpc systems, which is wrong.  This patch fixes it.

Reported-by: Martin Cermak <mcermak@redhat.com>
Reviewed-by: Andrea Cervesato <andrea.cervesato@suse.com>
Closes: https://github.com/linux-test-project/ltp/issues/1251