sigaltstack01: Fix check about alternative stack
Obviously condition for out of boundaries requires logical or, not
logical end.

Found when testing Gramine-SGX. sigaltstack01 mallocs a buffer with
size SIGSTKSZ (8192) and uses this buffer as an alternative stack.
However, in Gramine-SGX, it needs to reserve a large memory to save
xstate, sigframe, and restorer's address, thus it needs to reserve
memory larger than 11452 bytes, which overflows SIGSTKSZ.

Link: https://lore.kernel.org/ltp/tencent_05DE07BA6C168B5D5A26E2E54EC881EB6509@qq.com/
Link: https://github.com/gramineproject/gramine/issues/2153
Fixes: 865695bbc8 ("restructuring work done, and adding about 400 testcases")
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Liheng Chen <791960492@qq.com>
[ pvorel: reword commit message ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>