ioctl09: Use proper device for partitioning
The test occasionally fails because loop0p2 partition never gets
created. This happens because the uses parted(8) to modify partitions in
the backing file (test.img) while the loop device with this file as a
backing is already setup. Thus the page cache of loop0 can cache old
version of the partition table which is then used by the partitioning
code. Fix the problem by calling parted against the loop device
directly.

Link: https://lore.kernel.org/ltp/20250829141932.31997-1-jack@suse.cz/
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Tested-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Jan Kara <jack@suse.cz>