cve/cve-2015-3290: Correct the length parameter for perf munmap
Since commit b024d7b56c77 in the Linux kernel, the perf munmap operation
no longer allows VMA splits. However, in the cve-2015-3290 test case, the
length parameters passed to mmap and munmap were different:

* The length for mmap was event_mlock_kb * 1024, where event_mlock_kb is
  typically 516.
* The length for munmap was hardcoded to 512 * 1024.

This discrepancy caused the munmap operation to attempt to split a VMA due
to unaligned range boundaries, which resulted in it returning -EINVAL and
consequently caused the test case to fail.

Fix this by correcting the length parameter for the perf munmap call to
event_mlock_kb * 1024.

Signed-off-by: Tengda Wu <stavewu@gmail.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>