lib/tst_test: Fix FS kernel supported check
In case that there is only single filesystem defined in
tst_test.filesystems we format the device in the setup rather in the
function that iterates over all filesystems. However we missed a check
if a filesystem is supported in this shortcut, hence if mkfs for the
filesystem was installed and kernel support was missing the test failed
when it attempted to mount the filesystem.

Link: https://lore.kernel.org/ltp/20251027150924.17679-1-chrubis@suse.cz/
Reported-by: Anders Roxell <anders.roxell@linaro.org>
Tested-by: Anders Roxell <anders.roxell@linaro.org>
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>