Modify part of the name parameter in shm_open in ltp.
The name parameter of shm_open needs to start with a forward slash:
The  operation of shm_open() is analogous to that of open(2).  name specifies the shared memory object to be created or opened.  For porta‐
ble use, a shared memory object should be identified by a name of the form /somename; that is, a null-terminated string of up  to  NAME_MAX
(i.e., 255) characters consisting of an initial slash, followed by one or more characters, none of which are slashes.

Signed-off-by: tengshuangshuang <tengshuangshuang@xiaomi.com>