ci: do not checkout submodules
This was originally added so that kirk installation is tested. The
problem is that this brings random failures when git.kernel.org timeouts
when we are checking out submodules that are not needed for the build
and neither for the installation.

Unfortunatelly there is no support for selective submodule installation
in github checkout action. There is an issue and pull request hanging:

https://github.com/actions/checkout/pull/1866
https://github.com/actions/checkout/issues/1969

It looks like if we wanted to checkout only kirk submodule we would have
to install git and run all the commands ourselves in the CI, which I
would like to avoid. Hence this commit removes the submodules checkout
from the CI, with that we sacrifice kirk installation testing but avoid
random failures.

Reverts: 9cde2740cc ("ci: fetch submodules during checkout")

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>