ci: fix patch download when using curl
curl started to fail on CI after bots have been introduced to verify
that a human is interacting with patchwork. This creates an issue when
we need to download patches, since at the very first redirection we will
miss the server hosting our patches. By using `--location-trusted`, we
trust the bots redirection and we make sure that it's possible to
download the patch which is going to be applied via `git am`.

Acked-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Li Wang <liwang@redhat.com>
Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>